Initial commit of code, with work in progress on method 3
The first two methods to try and solve the calcudoku are present, but commented out since they were not able to solve it. In the following commits, the code will be removed. I didn't do that before, because I did not have a repository to keep track of the changes. Now I do. :)