Write solution to file, simplify initial grid construction and add timing
1. Solution to file
It is now possible to write the solution to file. By default it will store the output in a folder named 'output' in the working directory. No option to set the output folder yet.

2. Simplify initial grid
The previous way to fill the grid worked fine for 4x4 puzzles, but not 5x5. Instead of searching for a valid starting grid, the grid is set to ones. This will increase the time required to find a solution, but is less error prone.

3. Add timing
The time it takes to solve a grid is recorded and printed when the solution is found.